Add a required patch to fix the CI failures for OVN deployments with ovs 2.8
With the present ovs 2.8.1 and OVN deployment (tripelo or packstack), there is an
issue where the VMs don't get the DHCP response from the ovn-controller. This issue
is seen with OVS 2.8.1 and kernel 3.10.0-693.2.2.el7.x86_64. The issue is not seen
with earlier kernel versions.

The patch 0001-Check-flow-s-dl_type-before-setting-ct_orig_tuple-in.patch fixes
this issue. This patch is merged upstream in master and branch 2.8. There will
be ovs 2.8.2 soon. This commit is a temporary until we have ovs 2.8.2 so that
we can have ovs 2.8.1 with this patch to unblock OVN deployments.

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>